refactor(Lotto): LottoTicket을 domain/ticket 패키지 안으로 이동
