+++ src/main/java/domain/LottoSystem.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.ticket.LottoTicket;
+
 import java.util.List;
 
 public class LottoSystem {
